Separate authenticated and non-auth tests
