+++ test/e2e/package.json
@@ -5,6 +5,7 @@
     "test": "playwright test --project=chromium",
     "test:install": "playwright install",
     "test:all": "playwright test --project=chromium --project=firefox # limit to Chromium and FF as webkit currently fails",
+    "test:nonauth": "SKIP_AUTH=true playwright test --project=chromium --project=firefox tests/*.spec.js",
     "test:ui": "playwright test --project=chromium --ui",
     "test:debug": "playwright test --project=chromium --debug",
     "test:report": "playwright show-report"
