+++ test/e2e/tests/authenticated/collab.spec.js
@@ -10,7 +10,7 @@
  * governing permissions and limitations under the License.
  */
 import { test, expect } from '@playwright/test';
-import { getTestPageURL } from '../utils/page.js';
+import { getTestPageURL } from '../../utils/page.js';
 
 test('Collab cursors in multiple editors', async ({ browser, page }, workerInfo) => {
   // Open 2 editors on the same page and edit in both of them. One editor is logged in,
